+++ src/handler/processors/dependecies.jl
@@ -1,0 +1,7 @@
+function process_dependencies(request::HTTP.Request, state::HandlerState)
+    return HTTP.Response(
+        200,
+        ["Content-Type" => "application/json"],
+        body = state.cache.dependencies_json
+        )
+end
