+++ test/Project.toml
@@ -7,7 +7,6 @@ JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
 JSON2 = "2535ab7d-5cd8-5a07-80ac-9b1792aadce3"
 MD5 = "6ac74813-4b46-53a4-afec-0b5dc9d7885c"
 LibGit2 = "76f85450-5226-5b5a-8eaa-529ad045b433"
-MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 PlotlyBase = "a03496cd-edff-5a9b-9e67-9cda94a718b5"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 SnoopCompile = "aa65fe97-06da-5843-b5b1-d5d13cad87d2"
