+++ test/integration/base/jl_simple_app/jlstr001_jl_with_string.jl
@@ -1,0 +1,10 @@
+using Dash
+using DashHtmlComponents
+
+app = dash()
+
+app.layout = html_div() do
+    html_div("Hello Dash.jl testing", id="container")
+end
+
+run_server(app)
