+++ test/integration/base/jl_test_meta/jltm001_test_meta.jl
@@ -1,0 +1,9 @@
+using Dash
+using DashHtmlComponents
+using DashCoreComponents
+
+app = dash(meta_tags = [Dict(["name"=>"description", "content" => "some content"])])
+
+app.layout = html_div(children = "Hello world!", id = "hello-div")
+
+run_server(app)
