+++ test/integration/base/jl_test_meta/jltm002_test_meta.jl
@@ -1,0 +1,8 @@
+using Dash
+using DashHtmlComponents
+
+app = dash(meta_tags = [Dict(["charset"=>"ISO-8859-1", "keywords"=>"dash,pleasant,productive", "http-equiv"=>"content-type", "content"=>"text/html"])])
+
+app.layout = html_div(children = "Hello world!", id = "hello-div")
+
+run_server(app)
